requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
cryptography
